requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
python-dotenv>=0.19.0

# Document processing
pytesseract>=0.3.8
pdf2image>=1.16.0
PyPDF2>=2.0.0
python-magic>=0.4.24

# AI/ML components
langchain>=0.0.200
openai>=0.27.0
transformers>=4.11.0
torch>=1.9.0
sentence-transformers>=2.2.0
numpy>=1.21.0
scikit-learn>=0.24.0

# Database
sqlalchemy>=1.4.0
aiosqlite>=0.17.0
alembic>=1.7.0

# Utilities
python-dateutil>=2.8.2
pandas>=1.3.0
nltk>=3.6.0
spacy>=3.1.0
tqdm>=4.62.0

# Testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
pytest-cov>=2.12.0

# Development
black>=21.7b0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910

# Enhanced features
plotly>=5.3.0
jinja2>=3.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0 